% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{ActualsAndScores}
\alias{ActualsAndScores}
\title{ActualsAndScores}
\format{A data frame with 170 rows and 2 variables}
\usage{
data(ActualsAndScores)
}
\description{
A dataset containing the actuals for a simulated binary response variable as a numeric
 and the prediction probablity scores for a classification model like logistic regression.
}
\details{
\itemize{
  \item Actuals. A simulated variable meant to serve as the actual binary response variable. The good/events are marked as 1 while the bads/non-events are marked 0.
  \item PredictedScores. The prediction probability scores based on a classification model.
}
}
\keyword{datasets}

