% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{fscore}
\alias{fscore}
\title{fscore}
\usage{
fscore(actuals, predictedScores, threshold = 0.5)
}
\arguments{
\item{actuals}{The actual binary flags for the response variable. It can take a numeric vector containing values of either 1 or 0, where 1 represents the 'Good' or 'Events' while 0 represents 'Bad' or 'Non-Events'.}

\item{predictedScores}{The prediction probability scores for each observation. If your classification model gives the 1/0 predcitions, convert it to a numeric vector of 1's and 0's.}

\item{threshold}{If predicted value is above the threshold, it will be considered as an event (1), else it will be a non-event (0). Defaults to 0.5.}
}
\value{
The F1 Score
}
\description{
Computes the F1 Score for a given pair of actuals and predicted scores
}
\details{
F1 Score is computed as a weighted average of precision and recall. F1 score reaches its best value at 1 and worst at 0.
F1 Score = (2 * recall * precision) / (recall + precision)
}
\examples{
data('ActualsAndScores')
fscore(actuals=ActualsAndScores$Actuals, predictedScores=ActualsAndScores$PredictedScores)
}
\author{
Selva Prabhakaran \email{selva86@gmail.com}
}

