% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{SimData}
\format{A data frame with 30000 rows and 2 variables}
\usage{
data(SimData)
}
\description{
A dataset containing the actuals for a simulated binary response variable (Y) as a numeric
 and a categorical X variable with 9 groups, for which WOE calculation is performed.
}
\details{
\itemize{
  \item Y.Binary. A simulated variable meant to serve as the actual binary response variable. The good/events are marked as 1 while the bads/non-events are marked 0.
  \item X.Cat. A categorical variable (factor) with 9 groups.
}
}
\keyword{datasets}

